move to genesis_test file
